feat: move AppendixEditor out to common components
DEVSU-1757
